+++ src/components/Checkbox/Label.ts
@@ -1,12 +1,11 @@
 import styled, { css } from 'styled-components';
 import type { CheckboxDimension } from '#src/components/Checkbox/CheckboxDimension';
-import { TYPOGRAPHY } from '#src/components/Typography';
 
 export const styleTextMixin = css<{ dimension: CheckboxDimension; disabled?: boolean }>`
   font-size: ${({ dimension }) => (dimension === 'm' ? 16 : 14)}px;
   line-height: ${({ dimension }) => (dimension === 'm' ? 20 : 16)}px;
   font-weight: normal;
-  font-family: ${TYPOGRAPHY.fontFamily};
+  font-family: ${(p) => p.theme.fontFamily};
   color: ${({ disabled, theme }) => (disabled ? theme.color['Neutral/Neutral 30'] : theme.color['Neutral/Neutral 90'])};
   user-select: none;
 `;